requirements.txt
torch==2.2.1+cu121 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.17.1+cu121 --index-url https://download.pytorch.org/whl/cu121
tqdm>=4.66.1
numpy>=1.26.0
pandas>=2.1.1
Pillow>=10.0.0
pyyaml>=6.0
scikit-learn>=1.3.0
matplotlib>=3.8.1
tensorboard>=2.15.0
opencv-python-headless>=4.8.0
albumentations>=1.3.0
torchmetrics>=0.11.2
